model_training/requirements.txt
torch
torchvision
timm
torchmetrics
numpy
Pillow
matplotlib